+++ src/Constants.cs
@@ -1,0 +1,8 @@
+﻿namespace TypeScriptDefinitionGenerator
+{
+    public static class Constants
+    {
+        public const string FileExtension = ".d.ts";
+        public static string[] SupportedSourceExtensions { get; } = { ".cs", ".vb" };
+    }
+}